refactor(js-sdk): name the predicate module for its contents
`brand.ts` was named after the technique (a "brand check" in spec terms),
which reads as branded/nominal types in a TypeScript codebase and tells
you nothing at the call site. `is.ts` matches what the file holds and how
it is imported — `isBlobLike` from `./is` — and follows `errors.ts` in
naming the category rather than the mechanism.

Also notes @sindresorhus/is as prior art for the toStringTag one-liner,
and why it isn't a dependency: no Request or ReadableStream check.
